pyproject.toml
[tool.ruff]
line-length = 120
target-version = 'py312'

[tool.pytest.ini_options]
minversion = "8.3"
addopts = "-ra -q"
testpaths = [
    "backend/tests/"
]

[tool.poetry]
name = "scout"
version = "0.1.0"
description = "IPA insights engine"
authors = ["i.AI <i-dot-ai-enquiries@cabinetoffice.gov.uk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10, <3.13"
boto3 = "1.34.84"
sqlalchemy = ">=2.0.32"
pydantic = ">=2.8.2"
python-dotenv = "1.0.1"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
pydantic-settings = "^2.4.0"
tiktoken = "0.7.0"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.2"
decorator = "^5.1.1"
pre-commit = "^3.8.0"
yarl = "^1.11.1"
torch = "~2.2"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
deptry = "^0.16.2"
langchain = "0.1.20"
langchain-community = "0.0.38"
langchain-core = "0.1.52"
langchain-openai = "0.1.6"
numpy = "1.26.4"
regex = "2023.12.25"
unstructured = {extras = ["all-docs"], version = "^0.14.6"}
presidio-analyzer = "^2.2.354"
presidio-anonymizer = "^2.2.354"
flashrank = "^0.2.6"
instructor = "^1.3.4"
mammoth = "^1.8.0"
chromadb = "^0.5.3"
cruft = "^2.15.0"
bertopic = "^0.16.3"
hdbscan = "^0.8.38.post1"
ruff = "^0.6.8"
pytest = "^8.3.3"
pytest-env = "^1.1.5"
pytest-mock = "^3.14.0"
pytest-cov = "^5.0.0"
pytest-dotenv = "^0.5.2"
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
